% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bison_brm.R
\name{bison_brm}
\alias{bison_brm}
\title{Fit a brms model on BISoN network data}
\usage{
bison_brm(formula, edgemodel, data, num_draws = 100, ...)
}
\arguments{
\item{formula}{A formula compatible with a brms model}

\item{edgemodel}{A fitted bisonR edge model}

\item{data}{A dataframe of regression variables compatible with a brms model}

\item{num_draws}{Number of draws from the network posterior to use when fitting model}

\item{...}{}

\item{family}{Regression family compatible with a brms model}
}
\value{
A fitted brms model
}
\description{
Fit a brms model on BISoN network data
}
