% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bison_brm.R
\name{bison_brm_get_prior}
\alias{bison_brm_get_prior}
\title{Get priors for BISoN brm model}
\usage{
bison_brm_get_prior(formula, edgemodel_list, data_list, ...)
}
\arguments{
\item{formula}{A formula compatible with a brms model}

\item{edgemodel_list}{A bisonR edge model (or a list of them)}

\item{data_list}{A dataframe of regression variables compatible with a brms model (or a list of them)}

\item{...}{Additional parameters to be passed to brms::get_prior()}
}
\value{
The output from brms::get_prior()
}
\description{
Wrapper for brms::get_prior() for use with bison_brm.
}
