% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bison_brm.R
\name{bison_mice}
\alias{bison_mice}
\title{Generate a mice mids object}
\usage{
bison_mice(
  edgemodel_list,
  data_list,
  param_names,
  target_name,
  metric_name,
  num_draws = 100
)
}
\arguments{
\item{edgemodel_list}{A bisonR edge model (or a list of them)}

\item{data_list}{A dataframe of regression variables compatible with a brms model (or a list of them)}

\item{num_draws}{Number of draws from the network posterior}

\item{formula}{Formula of regression model}
}
\value{
A mice mids object
}
\description{
Generate a mice mids object
}
