% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bison_mixture.R
\name{bison_mixture}
\alias{bison_mixture}
\title{Fit a dyadic mixture model}
\usage{
bison_mixture(edgemodel, num_components = 5, criterion = "ICL", verbose = TRUE)
}
\arguments{
\item{edgemodel}{A fitted edge weight model.}

\item{num_components}{Maximum number of components to fit.}

\item{criterion}{"ICL" or "BIC" describing the criterion to use for model comparison.}

\item{verbose}{TRUE/FALSE indicating whether to output messages and progress when fitting.}
}
\value{
A bison_mixture object
}
\description{
Fit a dyadic mixture model
}
