% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bison_model.R
\name{bison_model}
\alias{bison_model}
\title{Fit an edge model to data}
\usage{
bison_model(
  formula,
  data,
  data_type = c("binary", "count"),
  directed = FALSE,
  partial_pooling = FALSE,
  zero_inflated = FALSE,
  priors = NULL,
  refresh = 0,
  mc_cores = 4,
  iter_sampling = 1000,
  iter_warmup = 1000,
  priors_only = FALSE
)
}
\arguments{
\item{formula}{Formula specifying social events and sampling effort on the LHS and edge weights, fixed, and random effects on the RHS.}

\item{data}{Aggregated or disaggregated dataframe of dyadic observations.}

\item{data_type}{"binary", "count", or "duration", specifying the type of edge weight model to use.}

\item{directed}{\code{TRUE} or \code{FALSE} specifying whether the network is directed or not.}

\item{partial_pooling}{Whether to pool edge weights so that information is shared between edges.}

\item{zero_inflated}{Whether to use a zero-inflated model to model excess zeroes.}

\item{priors}{List of priors in the format supplied by \code{get_default_priors()}.}

\item{refresh}{Frequency of messages printed while running the sampler.}

\item{mc_cores}{Number of cores to use when running the sampler.}

\item{iter_sampling}{Number of iterations to use for posterior samples.}

\item{iter_warmup}{Number of iterations to use for warmup (will not be used for samples).}

\item{priors_only}{Whether to use priors as posteriors or to allow the posteriors to be updated by data.}
}
\value{
An S3 edge model object containing edge samples and processed data.
}
\description{
Fit an edge model to data
}
\details{
Fits a BISoN edge weight model to a user-provided dataframe. The function supports either aggregated (at the
dyad-level) or disaggregated (at the observation-level) dataframes. Node names or IDs need to be formatted
as factors with the same levels.

The type of edge model and the interpretation of edge weights used depends on \code{data_type}, and will change
the interpretation of the edge weights.
}
