% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bison_to_igraph}
\alias{bison_to_igraph}
\title{Convert bisonR edgemodel posterior to igraph network objects}
\usage{
bison_to_igraph(obj, num_draws = 1000)
}
\arguments{
\item{obj}{bisonR edge model}

\item{num_draws}{Number of draws to sample from network posterior}
}
\value{
List of igraph objects representing posterior distribution of networks
}
\description{
Convert bisonR edgemodel posterior to igraph network objects
}
