% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_duration_to_binary}
\alias{convert_duration_to_binary}
\title{Convert duration data to binary}
\usage{
convert_duration_to_binary(df, event_name, divisor_name, max_event_length)
}
\arguments{
\item{df}{Original dataframe, ready for use with \code{bison_model}.}

\item{event_name}{The variable name in the dataframe corresponding to event durations.}

\item{divisor_name}{The variable name in the dataframe corresponding to divisor durations (e.g. observation time).}

\item{max_event_length}{The maximum length of time a single social event could take place for.}
}
\value{
A dataframe the same as the input but with \code{event_name} and \code{divisor_name} columns replaced with values compatible with binary models.
}
\description{
Uses a naive method for converting duration data to binary data. Use with caution. Ensure that all time periods are in the same units.
}
