% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_gbi_to_bison}
\alias{convert_gbi_to_bison}
\title{Convert group-by-individual matrix to bisonR format}
\usage{
convert_gbi_to_bison(
  gbi,
  group_properties = NULL,
  individual_properties = NULL,
  individual_constraints = NULL
)
}
\arguments{
\item{gbi}{Group-by-individual (R x C) matrix corresponding to R groups of C individuals.}

\item{group_properties}{A vector of length R describing properties of the groups to be added to the dataframe.}

\item{individual_properties}{A vector of length C describing properties of the individuals to be added to the dataframe.}

\item{individual_constraints}{A vector of length C describing constraints of the individuals.}
}
\value{
A dataframe for use in bisonR, where each row corresponds to a possible social event between two individuals.
}
\description{
Convert group-by-individual matrix to bisonR format
}
