% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_model.R
\name{diagnostic_plot}
\alias{diagnostic_plot}
\title{Diagnostic plot for an edge model object}
\usage{
diagnostic_plot(obj, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Diagnostic plot for an edge model object
}
