% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_model.R
\name{draw_edgelist_samples}
\alias{draw_edgelist_samples}
\title{Draw samples from edgelist posterior}
\usage{
draw_edgelist_samples(obj, num_draws)
}
\arguments{
\item{num_draws}{}
}
\value{

}
\description{
Draw samples from edgelist posterior
}
