% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{draw_network_metric_samples}
\alias{draw_network_metric_samples}
\title{Draw samples from posterior network-level metrics for a fitted edge weight model}
\usage{
draw_network_metric_samples(
  obj,
  metric_name,
  num_draws = 1000,
  standardise = FALSE
)
}
\arguments{
\item{obj}{Fitted S3 edge weight model.}

\item{metric_name}{Character name of network-level metric function.}

\item{num_draws}{Number of posterior draws to use.}

\item{standardise}{TRUE/FALSE indicating whether to mean-center the metrics within sample.}
}
\value{
A matrix of metric samples where each column corresponds to a node, and each row corresponds to a posterior draw.
}
\description{
Draw samples from posterior network-level metrics for a fitted edge weight model
}
