% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyadic_regression.R
\name{dyadic_regression}
\alias{dyadic_regression}
\title{Fit dyadic regression model}
\usage{
dyadic_regression(formula, edgemodel, df, mc_cores = 4)
}
\arguments{
\item{mc_cores}{}
}
\value{

}
\description{
Fit dyadic regression model
}
