% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_mixture.R
\name{edge_mixture}
\alias{edge_mixture}
\title{Fit a dyadic mixture model}
\usage{
edge_mixture(edgemodel, num_components = 5, verbose = TRUE)
}
\arguments{
\item{edgemodel}{A fitted edge weight model}

\item{component_range}{Vector of length 2 describing the minimum and maximum number of components to try to fit}
}
\value{
A edge_mixture object
}
\description{
Fit a dyadic mixture model
}
