% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_model.R
\name{edge_model}
\alias{edge_model}
\title{Fit an edge model to data}
\usage{
edge_model(
  formula,
  data,
  data_type = c("binary", "count", "duration"),
  directed = FALSE,
  durations = 1,
  method = c("mcmc", "vb", "inla"),
  verbose = FALSE,
  mc_cores = 1
)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
Fit an edge model to data
}
