% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{extract_metric}
\alias{extract_metric}
\title{Extract samples from posterior metrics for a fitted edge weight model}
\usage{
extract_metric(obj, metric_string, num_draws = 1000, standardise = FALSE)
}
\arguments{
\item{obj}{Fitted S3 edge weight model.}

\item{metric_string}{Character name of metric function (e.g. node_strength).}

\item{num_draws}{Number of posterior draws to use.}

\item{standardise}{TRUE/FALSE indicating whether to mean-center the metrics within sample.}
}
\value{
A matrix of metric samples where each column corresponds to a node, and each row corresponds to a posterior draw.
}
\description{
Extract samples from posterior metrics for a fitted edge weight model
}
