% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_mixture.R
\name{get_component_means}
\alias{get_component_means}
\title{Get component means from edge mixture model}
\usage{
get_component_means(object, num_components, ci = 0.9)
}
\arguments{
\item{object}{An S3 edge mixture model object}

\item{num_components}{The number of components in the mixture model}

\item{ci}{Credible interval for component mean estimates}
}
\value{
A dataframe describing the posteriors of the component means
}
\description{
Get component means from edge mixture model
}
