% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_contrasts}
\alias{get_contrasts}
\title{Calculate contrasts between parameters}
\usage{
get_contrasts(obj, parameter_1, parameter_2, ci = 0.9)
}
\arguments{
\item{obj}{}
}
\value{

}
\description{
Calculate contrasts between parameters
}
