% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_mixture.R
\name{get_edge_component_probabilities}
\alias{get_edge_component_probabilities}
\title{Get probabilities of component membership for each edge}
\usage{
get_edge_component_probabilities(object, num_components)
}
\arguments{
\item{object}{An S3 edge mixture model object}

\item{num_components}{The number of components in the mixture model}
}
\value{
Dataframe summarising edge component probabilities for each edge
}
\description{
Get probabilities of component membership for each edge
}
