% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bison_model.R
\name{get_edgelist}
\alias{get_edgelist}
\title{Retrieves an edgelist with uncertainty for a fitted edge weight model object}
\usage{
get_edgelist(obj, ci = 0.9, transform = TRUE)
}
\arguments{
\item{obj}{An S3 edge model object.}

\item{ci}{Credible interval to use in summary, based on quantiles.}

\item{transform}{\code{TRUE} or \code{FALSE} specifying whether to transform the edge weights from the internal link function scale.}
}
\value{
A \code{data.frame} object with columns of node IDs, median, lower, and upper bounds.
}
\description{
Retrieves an edgelist with uncertainty for a fitted edge weight model object
}
