% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_mixture.R
\name{get_network_component_probabilities}
\alias{get_network_component_probabilities}
\title{Get probabilities of component membership for the entire network}
\usage{
get_network_component_probabilities(object)
}
\arguments{
\item{object}{An S3 edge mixture model object}
}
\value{
A dataframe of probabilities of numbers of components over the entire network
}
\description{
Get probabilities of component membership for the entire network
}
