% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_comparison.R
\name{model_comparison}
\alias{model_comparison}
\title{Run leave-one-out model comparison}
\usage{
model_comparison(models)
}
\arguments{
\item{models}{List of fitted models to compare}
}
\value{
The output of the \code{loo_compare} function for the fitted models.
}
\description{
Run leave-one-out model comparison
}
