% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodal_regression.R
\name{nodal_regression}
\alias{nodal_regression}
\title{Fit a nodel regression model}
\usage{
nodal_regression(
  formula,
  edgemodel,
  df,
  mc_cores = 4,
  refresh = 500,
  priors = NULL,
  priors_only = FALSE
)
}
\arguments{
\item{formula}{Formula of the model}

\item{edgemodel}{Fitted edge weight model}

\item{df}{Dataframe for regression}

\item{mc_cores}{Number of cores to use for the MCMC sampler}

\item{refresh}{Frequency of print-outs from MCMC sampler}

\item{priors}{List of priors in the format supplied by \code{get_default_priors()}.}

\item{priors_only}{Whether to use priors as posteriors or to allow the posteriors to be updated by data.}
}
\value{
An S3 nodal model object containing chain samples and processed data.
}
\description{
Fit a nodel regression model
}
