% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{plot.metric_samples}
\alias{plot.metric_samples}
\title{Plot metric samples}
\usage{
\method{plot}{metric_samples}(x, ...)
}
\arguments{
\item{x}{Network metric samples object}

\item{...}{Additional parameters to be passed to plot function}
}
\description{
Plot metric samples
}
