% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{plot.network_metric_samples}
\alias{plot.network_metric_samples}
\title{Plot network metric samples}
\usage{
\method{plot}{network_metric_samples}(x)
}
\arguments{
\item{x}{Network metric samples object}
}
\description{
Plot network metric samples
}
