% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{plot_metric}
\alias{plot_metric}
\title{Plot metric samples}
\usage{
plot_metric(x, ...)
}
\arguments{
\item{x}{Network metric samples object}

\item{...}{Additional parameters to be passed to plot function}
}
\description{
Plot metric samples
}
