% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_predictions}
\alias{plot_predictions}
\title{Posterior predictive checks}
\usage{
plot_predictions(obj, ...)
}
\arguments{
\item{obj}{A fitted S3 object for edge weight, dyadic regression, or nodal regression models.}

\item{...}{Additional arguments to be passed to predictive plots.}
}
\description{
Posterior predictive checks
}
\details{
Plots the densities of summary statistics of observed data against predictions made by a model. These plots
can be used to check the predictive performance of a model.
}
