% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyadic_regression.R
\name{print.dyadic_model}
\alias{print.dyadic_model}
\title{Print information about a fitted dyadic regression model}
\usage{
\method{print}{dyadic_model}(x, ci = 0.9, ...)
}
\arguments{
\item{x}{An S3 dyadic regression model}

\item{ci}{Credible interval to use in summary, based on quantiles.}

\item{...}{Additional arguments to be passed to summary calculations.}
}
\description{
Print information about a fitted dyadic regression model
}
