% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{print.network_metric_samples}
\alias{print.network_metric_samples}
\title{Print a network metric samples object}
\usage{
\method{print}{network_metric_samples}(x, ci = 0.9, ...)
}
\arguments{
\item{x}{An S3 network metric samples object}

\item{ci}{Credible interval to use for summary}

\item{...}{Additional parameters to be passed to print}
}
\description{
Print a network metric samples object
}
