% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodal_regression.R
\name{print.nodal_model}
\alias{print.nodal_model}
\title{Print information about a fitted nodal regression model}
\usage{
\method{print}{nodal_model}(x, ci = 0.9, ...)
}
\arguments{
\item{x}{An S3 nodal regression model}

\item{ci}{Credible interval to use in summary, based on quantiles.}

\item{...}{Additional arguments to be passed to summary calculations.}
}
\description{
Print information about a fitted nodal regression model
}
