% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bison_mixture.R
\name{print.summary.bison_mixture}
\alias{print.summary.bison_mixture}
\title{Print information about an edge mixture model}
\usage{
\method{print}{summary.bison_mixture}(x, digits = 3, ...)
}
\arguments{
\item{x}{An S3 edge mixture model summary.}

\item{digits}{Number of digits to round summary values to.}

\item{...}{Additional arguments to be passed to print functions.}
}
\description{
Print information about an edge mixture model
}
