% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bison_model.R
\name{print.summary.bison_model}
\alias{print.summary.bison_model}
\title{Prints out an edge model summary object.}
\usage{
\method{print}{summary.bison_model}(x, ...)
}
\arguments{
\item{x}{An S3 edge model object.}

\item{...}{Additional parameters to be passed to print function.}
}
\description{
Prints out an edge model summary object.
}
