% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyadic_regression.R
\name{print.summary.dyadic_model}
\alias{print.summary.dyadic_model}
\title{Print information about a dyadic regression model}
\usage{
\method{print}{summary.dyadic_model}(x, digits = 3, ...)
}
\arguments{
\item{x}{An S3 dyadic regression model}

\item{digits}{Number of digits for rounding coefficients.}

\item{...}{Additional arguments to be passed to summary calculations.}
}
\description{
Print information about a dyadic regression model
}
