% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_mixture.R
\name{print.summary.edge_mixture}
\alias{print.summary.edge_mixture}
\title{Print information about an edge mixture model}
\usage{
\method{print}{summary.edge_mixture}(x, digits = 3, ...)
}
\arguments{
\item{x}{An S3 edge mixture model summary}

\item{...}{Additional arguments to be passed to print functions.}
}
\description{
Print information about an edge mixture model
}
