% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodal_regression.R
\name{print.summary.nodal_model}
\alias{print.summary.nodal_model}
\title{Print information about a fitted nodal regression model}
\usage{
\method{print}{summary.nodal_model}(x, digits = 3, ...)
}
\arguments{
\item{x}{An S3 nodal regression model}

\item{digits}{Number of digits for rounding coefficients.}

\item{...}{Additional arguments to be passed to summary calculations.}
}
\description{
Print information about a fitted nodal regression model
}
