% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{prior_check}
\alias{prior_check}
\title{Prior checks}
\usage{
prior_check(priors, model_type, type = "density")
}
\arguments{
\item{priors}{List of priors for a model, can be retrieved using \code{get_default_prior()}.}

\item{model_type}{Type of model the priors will be used for (same as the argument for \code{get_default_prior()}).}

\item{type}{Type of prior check to run, \code{"value"} or \code{"prediction"}. Details below.}
}
\description{
Prior checks
}
\details{
The parameter \code{type} determines what type of prior check to run. \code{type="value"} will
plot the prior probability over the parameter value on the original scale. \code{type="predictive"}
will run a prior predictive plot, where predictions from the model are generated using only
prior probabilities (the model when not updated from the data).
}
