% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{simulate_edge_model}
\alias{simulate_edge_model}
\title{Simulates data}
\usage{
simulate_edge_model(
  model_type,
  aggregated,
  location_effect = TRUE,
  age_diff_effect = TRUE
)
}
\arguments{
\item{model_type}{Type of data to simulate, one of "binary", "count", and "duration".}

\item{aggregated}{TRUE/FALSE indicating whether to aggregate over dyads or not.}

\item{location_effect}{TRUE/FALSE indicating whether there is an effect of location on observations.}

\item{age_diff_effect}{TRUE/FALSE indicating whether there is an effect of age difference on observations.}
}
\value{
A dataframe of the format accepted by edge weight models.
}
\description{
Simulates data
}
