% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{simulate_edge_model_mixture}
\alias{simulate_edge_model_mixture}
\title{Simulate data for an edge model with mixture components}
\usage{
simulate_edge_model_mixture(model_type, num_components, component_weights)
}
\arguments{
\item{model_type}{Type of data to simulate, one of "binary", "count", and "duration".}

\item{num_components}{Number of mixture components in the simulation.}

\item{component_weights}{Vector of component weightings for each component. Must sum to 1.}
}
\value{
A dataframe of the format accepted by edge weight models.
}
\description{
Simulate data for an edge model with mixture components
}
