% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bison_model.R
\name{summary.bison_model}
\alias{summary.bison_model}
\title{Generates a summary object for an edge model object}
\usage{
\method{summary}{bison_model}(object, ci = 0.9, transform = TRUE, ...)
}
\arguments{
\item{object}{An S3 edge model object to be summarised.}

\item{ci}{Credible interval to use in summary, based on quantiles.}

\item{transform}{\code{TRUE} or \code{FALSE} specifying whether to transform the edge weights from the internal link function scale.}

\item{...}{Additional arguments}
}
\description{
Generates a summary object for an edge model object
}
