% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyadic_regression.R
\name{summary.dyadic_model}
\alias{summary.dyadic_model}
\title{Summary of a fitted dyadic regression model}
\usage{
\method{summary}{dyadic_model}(object, ...)
}
\arguments{
\item{object}{An S3 dyadic regression model}

\item{...}{Additional arguments to be passed to summary calculations.}
}
\description{
Summary of a fitted dyadic regression model
}
