% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_model.R
\name{summary.edge_model}
\alias{summary.edge_model}
\title{Summarises a fitted edge model object}
\usage{
\method{summary}{edge_model}(object, ...)
}
\arguments{
\item{obj}{}
}
\value{

}
\description{
Summarises a fitted edge model object
}
