% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bison_brm.R
\name{bison_brm}
\alias{bison_brm}
\title{Fit a brms model on BISoN network data}
\usage{
bison_brm(formula, edgemodel_list, data_list, num_draws = 100, ...)
}
\arguments{
\item{formula}{A formula compatible with a brms model}

\item{edgemodel_list}{A bisonR edge model (or a list of them)}

\item{data_list}{A dataframe of regression variables compatible with a brms model (or a list of them)}

\item{num_draws}{Number of draws from the network posterior to use when fitting model}

\item{...}{}
}
\value{
A fitted brms model
}
\description{
Fit a brms model on BISoN network data
}
