% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bison_mice.R
\name{bison_mice}
\alias{bison_mice}
\title{Generate a mice mids object}
\usage{
bison_mice(
  edgemodel_list,
  data_list,
  param_names,
  target_name,
  metric_name,
  num_draws = 100,
  z_score = FALSE
)
}
\arguments{
\item{edgemodel_list}{A bisonR edge model (or a list of them)}

\item{data_list}{A dataframe of regression variables compatible with a brms model (or a list of them).}

\item{param_names}{List of column names for network features (e.g. node_1_id, node_2_id or just node_id)}

\item{target_name}{Name of network feature (edge, node, global)}

\item{metric_name}{Name of network metric to calculate}

\item{num_draws}{Number of draws from the network posterior}

\item{z_score}{Whether to Z-score bison variable or not.}
}
\value{
A mice mids object
}
\description{
Generate a mice mids object
}
