% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bisonR-package.R
\docType{package}
\name{bisonR-package}
\alias{bisonR-package}
\alias{bisonR}
\title{The 'bisonR' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2021). RStan: the R interface to Stan. R package version 2.21.3. https://mc-stan.org
}
