% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_model.R
\name{draw_edgelist_samples}
\alias{draw_edgelist_samples}
\title{Draw samples from edgelist posterior}
\usage{
draw_edgelist_samples(obj, num_draws)
}
\arguments{
\item{obj}{An S3 edge model object to be summarised.}

\item{num_draws}{Number of sample distributions to draw from the posterior.}
}
\value{
A \code{data.frame} of sample draws from the posteriors, where each column corresponds to a posterior draw.
}
\description{
Draw samples from edgelist posterior
}
