% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_metrics.R
\name{draw_node_metric_samples}
\alias{draw_node_metric_samples}
\title{Draw samples from the posterior node metric for a fitted edge weight model}
\usage{
draw_node_metric_samples(obj, metric_name, num_draws = 1000)
}
\arguments{
\item{metric_name}{}
}
\value{

}
\description{
Draw samples from the posterior node metric for a fitted edge weight model
}
