% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyadic_regression.R
\name{dyadic_regression}
\alias{dyadic_regression}
\title{Fit dyadic regression model}
\usage{
dyadic_regression(
  formula,
  edgemodel,
  df,
  mc_cores = 4,
  refresh = 500,
  mm = TRUE,
  priors = NULL,
  priors_only = FALSE
)
}
\arguments{
\item{formula}{Formula of the model}

\item{edgemodel}{Fitted edge weight model}

\item{df}{Dataframe for regression}

\item{mc_cores}{Number of cores to use for the MCMC sampler}

\item{refresh}{Frequency of print-outs from MCMC sampler}

\item{mm}{TRUE/FALSE indicating whether to include multi-membership effects in the regression}

\item{priors}{List of priors in the format supplied by \code{get_default_priors()}.}

\item{priors_only}{Whether to use priors as posteriors or to allow the posteriors to be updated by data.}
}
\value{
An S3 dyadic model object containing chain samples and processed data.
}
\description{
Fit dyadic regression model
}
\details{
Fits a dyadic regression mixed model of the form where edge weight (with uncertainty) is either a response or predictor.
}
