% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyadic_regression.R
\name{dyadic_regression}
\alias{dyadic_regression}
\title{Fit dyadic regression model}
\usage{
dyadic_regression(
  formula,
  edgemodel,
  df,
  mc_cores = 4,
  refresh = 500,
  mm = TRUE
)
}
\arguments{
\item{formula}{Formula of the model}

\item{edgemodel}{Fitted edge weight model}

\item{df}{Dataframe for regression}

\item{mc_cores}{Number of cores to use for the MCMC sampler}

\item{refresh}{Frequency of print-outs from MCMC sampler}

\item{mm}{TRUE/FALSE indicating whether to include multi-membership effects in the regression}
}
\value{
Fitted dyadic regression model object
}
\description{
Fit dyadic regression model
}
