% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{get_default_priors}
\alias{get_default_priors}
\title{Return default priors for a given model}
\usage{
get_default_priors(model_type)
}
\arguments{
\item{model_type}{}
}
\value{

}
\description{
Return default priors for a given model
}
