% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_model.R
\name{get_edgelist}
\alias{get_edgelist}
\title{Retrieves an edgelist with uncertainty for a fitted edge weight model object}
\usage{
get_edgelist(obj, ci = 0.9)
}
\arguments{
\item{obj}{}
}
\value{

}
\description{
Retrieves an edgelist with uncertainty for a fitted edge weight model object
}
