% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodal_regression.R
\name{nodal_regression}
\alias{nodal_regression}
\title{Fit a nodel regression model}
\usage{
nodal_regression(formula, edgemodel, df, mc_cores = 4)
}
\arguments{
\item{mc_cores}{}
}
\value{

}
\description{
Fit a nodel regression model
}
