% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_model.R
\name{plot_network}
\alias{plot_network}
\title{Sociogram plot with uncertainty of a fitted edge weight model object}
\usage{
plot_network(obj, ci = 0.9, lwd = 1, ciwd = 10)
}
\arguments{
\item{obj}{}
}
\value{

}
\description{
Sociogram plot with uncertainty of a fitted edge weight model object
}
