% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_model.R
\name{plot_trace}
\alias{plot_trace}
\title{Trace plot for a fitted edge model object}
\usage{
plot_trace(obj, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Trace plot for a fitted edge model object
}
