% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_trace}
\alias{plot_trace}
\title{Traceplot of MCMC chains}
\usage{
plot_trace(obj, ...)
}
\arguments{
\item{obj}{A fitted S3 object for edge weight, dyadic regression, or nodal regression models.}

\item{...}{Additional arguments to be passed to trace plots.}
}
\description{
Traceplot of MCMC chains
}
\details{
Plots MCMC chains for a fitted model. If the model has too many parameters to plot, only the first 12
will be plotted. To change the parameters being plotted, pass the \code{par_ids} argument
}
