% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_trace}
\alias{plot_trace}
\title{Traceplot of MCMC chains}
\usage{
plot_trace(obj, par_ids = 1:5, ...)
}
\arguments{
\item{obj}{A fitted S3 object for edge weight, dyadic regression, or nodal regression models.}

\item{par_ids}{A vector of parameter IDs to be plotted.}

\item{...}{Additional arguments to be passed to base R plots.}
}
\description{
Traceplot of MCMC chains
}
\details{
Plots MCMC chains for a fitted model. If the model has too many parameters to plot, only the first 12
will be plotted. To change the parameters being plotted, pass the \code{par_ids} argument
}
