% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyadic_regression.R
\name{print.dyadic_model}
\alias{print.dyadic_model}
\title{Prints out details of a fitted dyadic regression model}
\usage{
\method{print}{dyadic_model}(obj)
}
\arguments{
\item{obj}{}
}
\value{

}
\description{
Prints out details of a fitted dyadic regression model
}
