% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_model.R
\name{print.edge_model}
\alias{print.edge_model}
\title{Prints out details of a fitted edge model object}
\usage{
\method{print}{edge_model}(object, ci = 0.9, transform = TRUE)
}
\arguments{
\item{ci}{}
}
\description{
Prints out details of a fitted edge model object
}
