% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{prior_predictive_check}
\alias{prior_predictive_check}
\title{Run prior predictive check for a model specification}
\usage{
prior_predictive_check(
  formula,
  data,
  model_type,
  directed = FALSE,
  priors = NULL
)
}
\arguments{
\item{formula}{Formula for the model}

\item{data}{Data for the model}

\item{model_type}{Type of model to fit}

\item{priors}{List of priors for the model}
}
\description{
Run prior predictive check for a model specification
}
