% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{simulate_edge_model}
\alias{simulate_edge_model}
\title{Simulates observations from an edge model}
\usage{
simulate_edge_model(model_type, aggregated)
}
\value{

}
\description{
Simulates observations from an edge model
}
