% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bison_mixture.R
\name{summary.bison_mixture}
\alias{summary.bison_mixture}
\title{Summary of a edge mixture model}
\usage{
\method{summary}{bison_mixture}(object, ...)
}
\arguments{
\item{object}{An S3 edge mixture model}

\item{...}{Additional arguments to be passed to summary calculations.}
}
\value{
Returns a summary object of an edge mixture model.
}
\description{
Summary of a edge mixture model
}
