% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodal_regression.R
\name{summary.nodal_model}
\alias{summary.nodal_model}
\title{Summary of a fitted nodal regression model}
\usage{
\method{summary}{nodal_model}(obj)
}
\arguments{
\item{obj}{}
}
\value{

}
\description{
Summary of a fitted nodal regression model
}
