% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodal_regression.R
\name{summary.nodal_model}
\alias{summary.nodal_model}
\title{Summary of a fitted nodal regression model}
\usage{
\method{summary}{nodal_model}(object, ...)
}
\arguments{
\item{object}{An S3 dyadic regression model}

\item{...}{Additional arguments to be passed to summary calculations.}
}
\description{
Summary of a fitted nodal regression model
}
